funding-db/requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
django-filter==2.2.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.4
